+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 0.1.3
+
+* Fix pip installion bug
+* Add a few new methods to `GP`:
+    * `dm_dtheta` -- compute the derivative of the mean of the
+      gaussian process with respect to its parameters
+	* `plot` -- use pyplot to plot the predictive mean and variance of
+      the gaussian process
+* Switch to semantic versioning
+
 ## Version 0.01.2
 
 * Add Sphinx support